+++ src/components/jobs/index.ts
@@ -2,3 +2,4 @@ export { default as JobsBanner } from "./JobsBanner";
 export { default as JobsCategories } from "./JobsCategories";
 export { default as JobsCompany } from "./JobsCompany";
 export { default as JobsList } from "./JobsList";
+export { default as JobsContainer } from "./JobsContainer";
