+++ src/components/shared/Footer.tsx
@@ -3,9 +3,10 @@
 import { Briefcase } from "lucide-react";
 import React from "react";
 import Link from "next/link";
+import { safeLocalStorage } from "@/lib/utils";
 
 const Footer = () => {
-  const user: any = JSON.parse(localStorage.getItem("user") || "{}");
+  const user: any = JSON.parse(safeLocalStorage.getItem("user") || "{}");
   return (
     <footer className="bg-gray-900 text-white py-16">
       <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
